rebase with feature/reorganize-per-qe-docs
