+++ pyproject.toml
@@ -17,7 +17,7 @@ Homepage = "https://github.com/nexient-llc/launch-cli"
 Issues = "https://github.com/nexient-llc/launch-cli/issues"
 
 [tool.setuptools]
-package-dir = {launch="launch"}
+package-dir = {launch="src/launch"}
 
 [project.optional-dependencies]
 dev = ["pytest~=7.0", "pytest-mock~=3.0", "responses~=0.24.0", "black>=23.11.0", "isort>=5.12.0"]
@@ -32,5 +32,5 @@ profile = "black"
 minversion = "7.0"
 addopts = "-ra"
 testpaths = [
-    "tests"
+    "test/unit"
 ]
