+++ src/launch/__init__.py
@@ -1,0 +1,7 @@
+from semver import Version
+
+VERSION = "0.1.1"
+
+SEMANTIC_VERSION = Version.parse(VERSION)
+GITHUB_ORG_NAME = "nexient-llc"
+GITHUB_REPO_NAME = "launch-cli"
